feat: add health check script and backtest runner
- Implemented a health check script to output system health status in JSON format.
- Created a backtest runner script that simulates trading strategies using historical data.
- Added a real-time bot script for paper trading with structured cycles.
- Introduced confluence strategy for AI proposal validation against technical indicators.
- Added unit and integration tests for various components including execution, risk management, and indicators.
- Included sample data for testing and ensured schema validation for AI actions.

+++ core/indicators.py
@@ -337,7 +337,7 @@ class IndicatorCalculator:
             bb = BollingerBands(
                 close=df['close'],
                 window=self.bb_period,
-                window_dev=int(self.bb_std),  # ta uses integer std dev multiplier
+                window_dev=float(self.bb_std),  # FIX: must stay float; int() truncated 1.8 -> 1
                 fillna=False
             )
